requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16
Pillow>=10.0.0

